python/requirements.txt
altgraph>=0.17.3
future>=0.18.3
macholib>=1.16.2
modulegraph>=0.19.5
py2app>=0.28.6
PyYAML>=6.0.1
tkintertable>=1.3.3
